backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
motor==3.3.2
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
bcrypt==4.1.2
python-swiftclient==4.4.0
python-keystoneclient==5.3.0
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
black==24.1.1
isort==5.13.2
mypy==1.8.0
pillow==10.2.0
python-magic==0.4.27
